+++ src/style/components/generics/_index.scss
@@ -7,7 +7,6 @@
 @use 'details';
 @use 'form';
 @use 'listing';
-@use 'listing_fetch_wrapper';
 @use 'navigator';
 @use 'notification';
 @use 'playlist_entry_minimal';
