+++ src/style/components/generics/listing/_fetch_wrapper.scss
