+++ src/style/components/generics/listing/_index.scss
@@ -1,0 +1,8 @@
+//
+// Dakara Project
+//
+// Listings main style file
+//
+
+@use 'listing';
+@use 'fetch_wrapper';
