+++ src/style/components/generics/listing/_listing.scss
